All changes made as described in the requirements
